+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
 
-        <activity android:name=".ui.SettingsActivity"></activity>
+        <activity android:name=".SettingsActivity"></activity>
         <activity android:name=".ManualCrop" />
   
         <meta-data
