+++ app/src/main/java/com/example/cropshot/SettingsActivity.java
@@ -1,4 +1,4 @@
-package com.example.cropshot.ui;
+package com.example.cropshot;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -6,9 +6,6 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
-
-import com.example.cropshot.MainActivity;
-import com.example.cropshot.R;
 
 public class SettingsActivity extends AppCompatActivity {
